lang-portal-django-react/backend/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
python-dotenv==1.0.1
sqlparse==0.5.3
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==4.1.0
